Refatoração Update
- Adicionado detalhes.
