+++ utilitarios/yo/element/tree/objeto/conexao/ConexaoUI.java
@@ -4,8 +4,8 @@ import java.awt.Graphics2D;
 import java.awt.geom.Ellipse2D;
 import java.awt.geom.Path2D;
 import java.awt.geom.Point2D;
-import element.tree.Tree;
-import element.tree.TreeUI;
+import element.tree.main.Tree;
+import element.tree.main.TreeUI;
 import element.tree.objeto.nodulo.Nodulo;
 import element.tree.objeto.conexao.segmento.Segmento;
 import element.tree.objeto.conexao.segmento.SegmentoST;
