+++ utilitarios/yo/element/tree/popup/Popup.java
@@ -19,9 +19,9 @@ import element.tree.propriedades.Borda;
 import element.tree.propriedades.Cor;
 import element.tree.propriedades.Grossura;
 import element.tree.propriedades.Icone;
-import element.tree.Tree;
-import element.tree.TreeST;
-import element.tree.TreeUI;
+import element.tree.main.Tree;
+import element.tree.main.TreeST;
+import element.tree.main.TreeUI;
 import element.tree.undoRedo.UndoRedoIcone;
 import element.tree.objeto.Objeto;
 import element.tree.objeto.conexao.Conexao;
