+++ utilitarios/yo/element/tree/popup/borda/BordaPick.java
@@ -6,9 +6,8 @@ import java.awt.event.MouseEvent;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-
 import element.tree.propriedades.Borda;
-import element.tree.Tree;
+import element.tree.main.Tree;
 import element.tree.objeto.Objeto;
 import element.tree.objeto.conexao.Conexao;
 import element.tree.objeto.modulo.Modulo;
