+++ utilitarios/yo/element/tree/popup/icon/IconePick.java
@@ -9,10 +9,9 @@ import java.awt.image.BufferedImage;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-
 import element.tree.propriedades.Icone;
 import element.tree.objeto.modulo.Modulo;
-import element.tree.Tree;
+import element.tree.main.Tree;
 import element.tree.popup.Item;
 import element.tree.popup.Popup;
 @SuppressWarnings("serial")
